requirements.txt
# =============================================================================
# QGAN-MNIST-Synthesis Requirements
# GPU-Accelerated Quantum GAN using PyTorch-native quantum simulation
# =============================================================================

# =============================================================================
# CORE: PyTorch (GPU Quantum Simulation + Classical Networks)
# =============================================================================
torch>=2.0.0
torchvision>=0.15.0

# =============================================================================
# DATA PROCESSING & UTILITIES
# =============================================================================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0             # For K-means clustering in mode coverage
scipy>=1.11.0                   # For FID calculation (linalg.sqrtm)

# =============================================================================
# VISUALIZATION
# =============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0

# =============================================================================
# EXPERIMENT UTILITIES
# =============================================================================
tqdm>=4.65.0                    # Progress bars
pyyaml>=6.0.0                   # YAML config parsing

# =============================================================================
# TESTING (optional, for development)
# =============================================================================
pytest>=7.0.0

# =============================================================================
# OPTIONAL: Uncomment if needed
# =============================================================================
# wandb>=0.15.0                 # Weights & Biases experiment tracking
# jupyter>=1.0.0                # Jupyter notebooks
# ipywidgets>=8.0.0             # Interactive widgets
